crewai/requirements.txt
# CrewAI Core
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# API et serveur
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenRouter / OpenAI
openai>=1.0.0
httpx>=0.25.0

# Recherche web
duckduckgo-search>=4.1.0

# Utilitaires
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Logging
loguru>=0.7.2

# Sécurité
cryptography>=41.0.0
